conecta-python/pyproject.toml
[build-system]
requires = ["maturin>=1.8,<2.0"]
build-backend = "maturin"

[project]
name = "conecta"
description = "Load data from SQL databases to Arrow"
readme = "README.md"
authors = [
    { name = "surister", email = "surister98@gmail.com"}
]
requires-python = ">=3.10"
keywords = ["arrow", "databases", "sql", "dataframe"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Rust",
    "Topic :: Scientific/Engineering",
    "Typing :: Typed",
]
dynamic = ["version"]
dependencies = [
]

[project.urls]
Homepage = "https://github.com/surister/conecta/"
Documentation = "https://conecta.surister.dev/"
Repository = "https://github.com/surister/conecta/"
Changelog = "https://github.com/surister/conecta/releases"

[tool.maturin]
features = ["pyo3/extension-module"]

[dependency-groups]
dev = [
    "arro3-core>=0.5.1",
    "black>=25.1.0",
    "nanoarrow>=0.7.0",
    "polars>=1.31.0",
    "psycopg2>=2.9.10",
    "pyarrow>=21.0.0",
    "sqlalchemy>=2.0.43",
    "testcontainers[postgres]>=4.13.1",
]
